chore: Start refactoring receiver lambda
Prepare for handling different types of notifications.
