+++ src/submission-forwarder/receiver-lambda/ErrorTypes.ts
@@ -1,0 +1,3 @@
+
+export class ParseError extends Error { }
+export class SendMessageError extends Error { }
